Fixed sorting issues

+++ src/app/components/home/home.component.ts
@@ -4,8 +4,8 @@ import { MatTableDataSource } from '@angular/material/table';
 import {MatDialog, MAT_DIALOG_DATA} from '@angular/material/dialog';
 import { AddBankComponent } from '../add-bank/add-bank.component';
 import { BankService } from 'src/app/services/bank.service';
-import { MatSort } from '@angular/material/sort';
 import { BankModel } from '../models/bank-model';
+import {MatSort, Sort} from '@angular/material/sort';
 
 
 @Component({
@@ -13,29 +13,36 @@ import { BankModel } from '../models/bank-model';
   templateUrl: './home.component.html',
   styleUrls: ['./home.component.css']
 })
-export class HomeComponent implements OnInit, AfterViewInit {
+export class HomeComponent implements OnInit {
 
   constructor(private dialog: MatDialog, private bankService: BankService) { }
 
+  @ViewChild(MatPaginator) paginator !: MatPaginator;
+  @ViewChild(MatSort) empTbSort = new MatSort();
+
+
+  displayedColumns: string[] = ['bankId', 'bankName', 'code', 'incorporateDate','noOfStaff','noOfBranches','status', 'action','filter','add'];
+  dataSource !: MatTableDataSource<any>;
   bankList : BankModel[] = [];
+
 
   ngOnInit(): void {
     this.getList();
   }
 
-  displayedColumns: string[] = ['bankId', 'bankName', 'code', 'incorporateDate','noOfStaff','noOfBranches','status', 'action','add'];
-  dataSource !: MatTableDataSource<any>;
+  applyFilter(event: Event) {
+    const filterValue = (event.target as HTMLInputElement).value;
+    this.dataSource.filter = filterValue.trim().toLowerCase();
 
-  @ViewChild(MatPaginator) paginator!: MatPaginator;
-  @ViewChild(MatSort) sort!: MatSort;
-
-  ngAfterViewInit() {
-    // this.dataSource.paginator = this.paginator;
+    if (this.dataSource.paginator) {
+      this.dataSource.paginator.firstPage();
+    }
   }
 
   openDialog() {
     this.dialog.open(AddBankComponent, {
-      width: '50%'
+      width: '50%',
+      data: ''
     });
   }
 
@@ -44,16 +51,19 @@ export class HomeComponent implements OnInit, AfterViewInit {
       // this.bankList = res;
       this.dataSource = new MatTableDataSource(res);
       this.dataSource.paginator = this.paginator;
-      this.dataSource.sort = this.sort;
+      this.dataSource.sort = this.empTbSort;
     });
   }
 
   onDelete(bankId: number){
-    this.bankService.deleteBank(bankId).subscribe(res => {
-      console.log(res);
-      console.log(bankId);
-      this.getList();
-    })
+    let isConfirm : boolean=confirm('Are you want to delete this Record?');
+
+    if(isConfirm){
+      this.bankService.deleteBank(bankId).subscribe(res => {
+        this.getList();
+      })
+    }
+
   }
 
   onUpdate(bank: BankModel){
@@ -65,27 +75,3 @@ export class HomeComponent implements OnInit, AfterViewInit {
 
 }
 
-
-
-// const ELEMENT_DATA: PeriodicElement[] = [
-//   {position: 1, name: 'Hydrogen', weight: 1.0079, symbol: 'H'},
-//   {position: 2, name: 'Helium', weight: 4.0026, symbol: 'He'},
-//   {position: 3, name: 'Lithium', weight: 6.941, symbol: 'Li'},
-//   {position: 4, name: 'Beryllium', weight: 9.0122, symbol: 'Be'},
-//   {position: 5, name: 'Boron', weight: 10.811, symbol: 'B'},
-//   {position: 6, name: 'Carbon', weight: 12.0107, symbol: 'C'},
-//   {position: 7, name: 'Nitrogen', weight: 14.0067, symbol: 'N'},
-//   {position: 8, name: 'Oxygen', weight: 15.9994, symbol: 'O'},
-//   {position: 9, name: 'Fluorine', weight: 18.9984, symbol: 'F'},
-//   {position: 10, name: 'Neon', weight: 20.1797, symbol: 'Ne'},
-//   {position: 11, name: 'Sodium', weight: 22.9897, symbol: 'Na'},
-//   {position: 12, name: 'Magnesium', weight: 24.305, symbol: 'Mg'},
-//   {position: 13, name: 'Aluminum', weight: 26.9815, symbol: 'Al'},
-//   {position: 14, name: 'Silicon', weight: 28.0855, symbol: 'Si'},
-//   {position: 15, name: 'Phosphorus', weight: 30.9738, symbol: 'P'},
-//   {position: 16, name: 'Sulfur', weight: 32.065, symbol: 'S'},
-//   {position: 17, name: 'Chlorine', weight: 35.453, symbol: 'Cl'},
-//   {position: 18, name: 'Argon', weight: 39.948, symbol: 'Ar'},
-//   {position: 19, name: 'Potassium', weight: 39.0983, symbol: 'K'},
-//   {position: 20, name: 'Calcium', weight: 40.078, symbol: 'Ca'},
-// ];
